+++ src/simulator.rs
@@ -11,25 +11,11 @@ pub use tracing::{
 
 use std::io::prelude::*;
 
-use crate::{node::Cell, universe::Universe};
-
-use sdl2::{
-    event::Event,
-    keyboard::Keycode,
-    pixels::Color,
-    rect::{Point, Rect},
-    render::Canvas,
-    video::Window,
-    EventPump,
-};
-
-const CELL_SIZE: u32 = 10;
-const CELL_PADDING: u32 = 2;
+use crate::{node::Cell, renderer::Renderer, universe::Universe};
 
 pub struct Simulator {
     universe: Universe,
-    canvas: Canvas<Window>,
-    event_pump: EventPump,
+    renderer: Renderer,
 }
 
 impl Simulator {
@@ -37,24 +23,9 @@ impl Simulator {
         let mut universe = Universe::new();
         universe.initalize();
 
-        // init sdl
-        let sdl_context = sdl2::init().unwrap();
-        let video_subsystem = sdl_context.video().unwrap();
+        let renderer = Renderer::new();
 
-        let window = video_subsystem
-            .window(env!("CARGO_PKG_NAME"), 1600, 1200)
-            .position_centered()
-            .build()
-            .unwrap();
-
-        let canvas = window.into_canvas().build().unwrap();
-        let event_pump = sdl_context.event_pump().unwrap();
-
-        Simulator {
-            universe,
-            canvas,
-            event_pump,
-        }
+        Simulator { universe, renderer }
     }
 
     pub fn run(&mut self) {
@@ -65,61 +36,12 @@ impl Simulator {
     }
 
     fn update(&mut self) {
-        for event in self.event_pump.poll_iter() {
-            match event {
-                Event::Quit { .. }
-                | Event::KeyDown {
-                    keycode: Some(Keycode::Escape),
-                    ..
-                }
-                | Event::KeyDown {
-                    keycode: Some(Keycode::Q),
-                    ..
-                } => std::process::exit(0),
-                _ => {}
-            }
-        }
-
         self.universe.evolve();
+        self.renderer.update();
     }
 
     pub fn render(&mut self) {
-        let canvas = &mut self.canvas;
-
-        canvas.set_draw_color(Color::BLACK);
-        canvas.clear();
-        canvas.set_draw_color(Color::WHITE);
-
-        let center = {
-            let size = canvas.viewport();
-            Point::new(size.width() as i32 / 2, size.height() as i32 / 2)
-        };
-
-        let x_range = center.x() / CELL_SIZE as i32;
-        let y_range = center.y() / CELL_SIZE as i32;
-
-        for y in -y_range..y_range {
-            for x in -x_range..x_range {
-                let alive = match self.universe.get_cell((x as i64, y as i64)) {
-                    Cell::Dead => false,
-                    Cell::Alive => true,
-                };
-
-                if alive {
-                    let center = center
-                        + Point::new(
-                            x as i32 * (CELL_SIZE + CELL_PADDING) as i32,
-                            y as i32 * (CELL_SIZE + CELL_PADDING) as i32,
-                        );
-                    let rect = Rect::from_center(center, CELL_SIZE, CELL_SIZE);
-
-                    canvas.fill_rect(rect).unwrap();
-                }
-            }
-        }
-
-        canvas.present();
-        //std::thread::sleep(std::time::Duration::from_millis(10));
+        self.renderer.render();
     }
 
     pub fn read_rls(&mut self, pattern: &str) {
